requirements.txt
pyTelegramBotAPI
google-generativeai
openai-whisper
python-dotenv
